# [SPARK-22025][PySpark] Speeding up fromInternal for StructField