--- a/python/pyspark/sql/types.py
+++ b/python/pyspark/sql/types.py
@@ -410,6 +410,9 @@ class StructField(DataType):
         self.dataType = dataType
         self.nullable = nullable
         self.metadata = metadata or {}
+        self.needConversion = dataType.needConversion
+        self.toInternal = dataType.toInternal
+        self.fromInternal = dataType.fromInternal
 
     def simpleString(self):
         return '%s:%s' % (self.name, self.dataType.simpleString())
@@ -430,15 +433,6 @@ class StructField(DataType):
                            _parse_datatype_json_value(json["type"]),
                            json["nullable"],
                            json["metadata"])
-
-    def needConversion(self):
-        return self.dataType.needConversion()
-
-    def toInternal(self, obj):
-        return self.dataType.toInternal(obj)
-
-    def fromInternal(self, obj):
-        return self.dataType.fromInternal(obj)
 
     def typeName(self):
         raise TypeError(