## What changes were proposed in this pull request?

Change function call to references can greatly speed up function calling.

Benchmark
```
df = spark.range(10000000).selectExpr("id as id0", "id as id1", "id as id2", "id as id3", "id as id4", "id as id5", "id as id6", "id as id7", "id as id8", "id as id9", "struct(id) as s").cache()
df.count()
df.rdd.map(lambda x: x).count()
```

Before:
```
420274688 function calls (410272560 primitive calls) in 1710.883 seconds

Ordered by: internal time, cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
110000000  412.988    0.000  648.632    0.000 types.py:442(fromInternal)
 10000000  250.461    0.000  899.093    0.000 types.py:629(<listcomp>)
 30000000  188.816    0.000 1409.060    0.000 types.py:622(fromInternal)
100000000  177.706    0.000  177.706    0.000 types.py:88(fromInternal)
```
After
```
310274584 function calls (300272456 primitive calls) in 1308.709 seconds

Ordered by: internal time, cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
 10000000  251.018    0.000  483.316    0.000 types.py:619(<listcomp>)
 30000000  190.364    0.000 1001.505    0.000 types.py:612(fromInternal)
100000000  175.242    0.000  175.242    0.000 types.py:88(fromInternal)
 20000000  155.427    0.000  325.382    0.000 types.py:1471(_create_row)
```

The change is 110 millions function calls less in types.py:442(fromInternal) and about 25% performance gain.

## How was this patch tested?

Existing tests.
Performance benchmark.