@dongjoon-hyun 2017-09-15
Hi, @maver1ck .
Could you add your benchmark result on PR description for reviewers? That will be your commit log finally.
> greatly speed up function calling