@HyukjinKwon 2017-09-20
I'd close this PR if there is no objection @maver1ck and I didn't miss something.