@maver1ck 2017-09-20
@HyukjinKwon
I created this before https://github.com/apache/spark/pull/19249, which greatly decrease function call.

I agree we can close it.