@HyukjinKwon 2017-09-20
Yea, thanks. That was a cool patch BTW.